% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACP.R
\name{correlationplot.UniNum}
\alias{correlationplot.UniNum}
\title{Correlation plot}
\usage{
correlationplot.UniNum(object, usedComp = 1)
}
\arguments{
\item{object}{of UniNum class}

\item{usedComp}{a float}
}
\value{

}
\description{
Charts a correlation matrix
}
\examples{
obj <- Perform_UniNum(data)
correlationplot.UniNum(obj,usedComp = 1)
}
