% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACP.R
\name{Perform_UniNum}
\alias{Perform_UniNum}
\title{Class Constructor}
\usage{
Perform_UniNum(data)
}
\arguments{
\item{data}{data frame}
}
\value{
\item{instance}{  object of UniNum class}
}
\description{
UniNum class creation
}
\examples{
dataset <- matrix(sample(c(1, 1:5), 25, replace= TRUE),5)
df <- as.data.frame(dataset)
Perform_UniNum(df)
}
