% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACP.R
\name{print.UniNum}
\alias{print.UniNum}
\title{print override}
\usage{
\method{print}{UniNum}(object)
}
\arguments{
\item{object}{UniNum instance}
}
\value{

}
\description{
Prints eigen values from the data of an UniNum object
}
\examples{
object <- Perform_UniNum(data)
print(object)
}
